+++ app/Filament/Resources/Products/RelationManagers/ImagesRelationManager.php
@@ -51,6 +51,7 @@ class ImagesRelationManager extends RelationManager
                 FileUpload::make('path')
                     ->label('Image')
                     ->image()
+                    ->maxSize(config('media.max_upload_size_kb'))
                     ->disk('public')
                     ->directory('product-images')
                     ->saveUploadedFileUsing(ConvertImageToWebp::forFileUpload())